Fix display of m of n fields and save state
